VOA-3213 [Technical Debt] RALD frontend - Remove usage of npm, grunt to compile CSS and Javascript
