+++ app/assets/javascripts/common.js
