+++ app/assets/javascripts/intelAlerts.js
