+++ app/assets/stylesheets/application2.scss
@@ -51,6 +51,3 @@ ul#header-account-info{
 @import "print_page_screen";
 @import "form";
 @import "spinner";
-
-
-
