+++ app/views/includes/javascripts.scala.html
@@ -35,13 +35,3 @@
 	<script src="@routes.Assets.at("javascripts/vendor/jsconf.min.js")" type="text/javascript"></script>
 	<script src="@routes.Assets.at("javascripts/polyfills/details.polyfill.min.js")" type="text/javascript"></script>
 	<script src="@routes.Assets.at("javascripts/polyfills/toISOString.polyfill.min.js")" type="text/javascript"></script>
-
-
-<!--[if lt IE 10]>
-	<script src="@routes.Assets.at("javascripts/vendor/jquery.placeholder.min.js")" type="text/javascript"></script>
-	<script type="text/javascript">
-		$(function() {
-	 		$('input, textarea').placeholder();
-		});
-	</script>
-<![endif]-->
