+++ build.sbt
@@ -1,4 +1,6 @@
 
+import com.typesafe.sbt.uglify.Import._
+import net.ground5hark.sbt.concat.Import._
 import play.core.PlayVersion
 import scoverage.ScoverageKeys
 import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin._
@@ -21,7 +23,6 @@ val scoverageSettings = {
 
 val compileDeps = Seq(
   filters,
-  ws,
   "uk.gov.hmrc" %% "bootstrap-frontend-play-28" % "5.24.0",
   "uk.gov.hmrc" %% "play-frontend-hmrc" % "3.20.0-play-28",
   "uk.gov.hmrc.mongo" %% "hmrc-mongo-play-28" % "0.64.0",
@@ -64,7 +65,7 @@ lazy val root = (project in file("."))
     routesGenerator := InjectedRoutesGenerator,
     Compile / unmanagedResourceDirectories += baseDirectory.value / "resources",
     majorVersion := 3,
-  ).settings(JavaScriptBuild.javaScriptUiSettings: _*)
+  )
   .configs(IntegrationTest)
   .settings(
     Defaults.itSettings,
@@ -72,3 +73,14 @@ lazy val root = (project in file("."))
   )
   .enablePlugins(PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin)
   .disablePlugins(plugins.JUnitXmlReportPlugin)
+  .settings(
+    Concat.groups := Seq(
+      "javascripts/app.js" -> group((baseDirectory.value / "app" / "assets" / "javascripts") * "*.js")
+    ),
+    // prevent removal of unused code which generates warning errors due to use of third-party libs
+    uglifyCompressOptions := Seq("unused=false", "dead_code=false"),
+    // below line required to force asset pipeline to operate in dev rather than only prod
+    Assets / pipelineStages := Seq(concat, uglify),
+    // only compress files generated by concat and files from polyfills directory
+    uglify / includeFilter := GlobFilter("app.js") || GlobFilter("*polyfill.js")
+  )
