--- package.json
@@ -1,47 +1,0 @@
-{
-  "name": "voa-framework",
-  "description": "VOA framework",
-  "version": "0.1.0",
-  "private": true,
-  "engines": {
-    "node": "~10.16.1"
-  },
-  "dependencies": {
-    "govuk-frontend": "~3.10.0",
-    "hmrc-frontend": "1.21.0"
-  },
-  "devDependencies": {
-    "grunt": "^1.1.0",
-    "grunt-cli": "^1.3.2",
-    "grunt-concurrent": "^3.0.0",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-contrib-connect": "^2.0.0",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-cssmin": "^3.0.0",
-    "grunt-contrib-jasmine": "^2.1.0",
-    "grunt-contrib-jshint": "^2.1.0",
-    "grunt-contrib-uglify": "^4.0.1",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-nodemon": "^0.4.2",
-    "grunt-sass": "^3.1.0",
-    "grunt-text-replace": "^0.4.0",
-    "node-sass": "^4.14.1"
-  },
-  "main": "Gruntfile.js",
-  "directories": {
-    "test": "test"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/hmrc/for-frontend.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/hmrc/for-frontend/issues"
-  },
-  "homepage": "https://github.com/hmrc/for-frontend#readme"
-}
