+++ project/plugins.sbt
@@ -12,3 +12,11 @@ addSbtPlugin("uk.gov.hmrc" % "sbt-auto-build" % "3.6.0")
 addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "2.4.0")
 
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.9.3")
+
+addSbtPlugin("io.github.irundaia" % "sbt-sassify" % "1.5.2")
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-jshint" % "1.0.6")
+
+addSbtPlugin("net.ground5hark.sbt" % "sbt-concat" % "0.2.0")
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-uglify" % "2.0.0")
